requirements.txt
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
category-encoders==2.6.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
deprecation==2.1.0
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
fonttools==4.51.0
greenlet==3.0.3
idna==3.7
imbalanced-learn==0.12.2
importlib_metadata==7.1.0
ipython==8.24.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaleido==0.2.1
kiwisolver==1.4.5
lightgbm==4.3.0
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mysqlclient==2.2.4
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.59.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.3
packaging==24.0
pandas==2.1.4
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycaret==3.3.2
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pyod==1.1.3
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
referencing==0.35.1
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.1
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
setuptools==69.5.1
six==1.16.0
sktime==0.26.0
stack-data==0.6.3
statsmodels==0.14.2
tbats==1.1.3
tenacity==8.3.0
threadpoolctl==3.5.0
tqdm==4.66.4
traitlets==5.14.3
tsdownsample==0.1.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
widgetsnbextension==4.0.10
xxhash==3.4.1
yellowbrick==1.5
zipp==3.18.2
zope.event==5.0
zope.interface==6.4.post2
